pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "err-x509"
version = "1.0.0"
description = "Fix Clash x509 SSL certificate errors by adding skip-cert-verify"
readme = "README.md"
authors = [
    {name = "13winged", email = ""}
]
license = {text = "MIT"}
keywords = ["clash", "ssl", "x509", "proxy", "yaml", "certificate"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: Proxy Servers",
    "Topic :: Security",
    "Topic :: Utilities",
]
requires-python = ">=3.7"
dependencies = [
    "pyyaml>=6.0",
    "click>=8.0",  # Для красивого CLI
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "flake8>=6.0",
    "mypy>=1.0",
    "twine>=4.0",
    "build>=0.10",
]

[project.urls]
Homepage = "https://github.com/13winged/err_x509"
Repository = "https://github.com/13winged/err_x509"
BugTracker = "https://github.com/13winged/err_x509/issues"
Documentation = "https://github.com/13winged/err_x509#readme"

[project.scripts]
err-x509 = "err_x509.cli:main"
err_x509 = "err_x509.cli:main"
fix-x509 = "err_x509.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["err_x509*"]

[tool.black]
line-length = 88
target-version = ['py37']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.7"
warn_return_any = true
warn_unused_configs = true